fix: Normalize uri both index and search
* Normalize uri before storing in index and before executing search
