+++ src/Spark.Engine/Search/ElementIndexer.cs
@@ -192,12 +192,7 @@ namespace Spark.Engine.Search
             if (element == null || String.Empty.Equals(element.Value))
                 return null;
 
-            if(Uri.TryCreate(element.Value, UriKind.RelativeOrAbsolute, out var uri))
-            {
-                return ListOf(new StringValue(uri.ToString()));
-            }
-
-            return ListOf(new StringValue(element.Value));
+            return ListOf(new StringValue(UriUtil.NormalizeUri(element.Value)));
         }
         private List<Expression> ToExpressions(Hl7.Fhir.Model.Date element)
         {
